requirement.txt
aniso8601==9.0.1
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
Flask==2.2.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.2.1
